# docs/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "CodeSage API",
    "description": "Local-first, stateless AI pull request review assistant.\n\n## Authentication\nNo user authentication. Configure `GitHub:PersonalAccessToken` and `OpenAI:ApiKey`\nvia user-secrets or environment variables.\n\n## Errors\nAll failures return RFC 7807 `application/problem+json` with an `errorCode` extension.\n\n## Versioning\nCurrent surface is `/api/v1`. Breaking changes will ship as `/api/v2`.",
    "contact": {
      "name": "CodeSage Contributors",
      "url": "https://github.com/codesage"
    },
    "license": {
      "name": "MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/reviews": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Create AI review",
        "description": "Runs the configured OpenAI model against a ReviewContext and returns a structured ReviewReport. Stateless — nothing is persisted. Obtain a ReviewContext from GET .../pull-requests/{number}/analysis.",
        "operationId": "CreateReviewV1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewContext"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewReport"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error — RFC 7807 ProblemDetails.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "traceId": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CodeLanguage": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9
        ],
        "type": "integer",
        "description": "Programming languages / formats CodeSage can classify in Stage 3.\nUnknown and Binary are explicit so future AI providers can filter safely.",
        "format": "int32"
      },
...
      "FileChangeStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "description": "Normalized change status — independent of any source-control vendor vocabulary.",
        "format": "int32"
      },
      "FindingCategory": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10,
          11,
          12
        ],
        "type": "integer",
        "description": "Category for an individual finding.",
        "format": "int32"
      },
      "FindingSeverity": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "description": "Severity for an individual finding.",
        "format": "int32"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ReviewAuthorInfo": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewCommitInfo": {
        "type": "object",
        "properties": {
          "sha": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "authorLogin": {
            "type": "string",
            "nullable": true
          },
          "committedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewContext": {
        "type": "object",
        "properties": {
          "repository": {
            "$ref": "#/components/schemas/ReviewRepositoryInfo"
          },
          "pullRequest": {
            "$ref": "#/components/schemas/ReviewPullRequestInfo"
          },
          "author": {
            "$ref": "#/components/schemas/ReviewAuthorInfo"
          },
          "commits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewCommitInfo"
            }
          },
          "changedFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewFileChange"
            }
          },
          "statistics": {
            "$ref": "#/components/schemas/ReviewStatistics"
          },
          "languageBreakdown": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "summary": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "The only object future AI providers should consume.\nContains no GitHub / Octokit / vendor types — only CodeSage vocabulary.",
        "example": "{\"repository\":{\"name\":\"Healthcare.API\",\"fullName\":\"acme/Healthcare.API\",\"defaultBranch\":\"main\"},\"pullRequest\":{\"number\":42,\"title\":\"Harden auth\",\"description\":\"...\",\"state\":\"open\",\"isDraft\":false,\"baseRef\":\"main\",\"headRef\":\"feature/auth\",\"createdAt\":\"2026-07-01T12:00:00Z\",\"updatedAt\":\"2026-07-02T09:00:00Z\"},\"author\":{\"login\":\"dev\",\"name\":null,\"avatarUrl\":null},\"commits\":[],\"changedFiles\":[],\"statistics\":{\"fileCount\":1,\"commitCount\":1,\"additions\":10,\"deletions\":2,\"totalChangedLines\":12,\"languagesUsed\":[\"C#\"],\"largestModifiedFile\":\"Auth.cs\",\"largestModifiedFileChanges\":12,\"testFilesChanged\":0,\"sqlFilesChanged\":0,\"configurationFilesChanged\":0,\"controllerFilesChanged\":0,\"serviceFilesChanged\":0,\"sqlModified\":false},\"languageBreakdown\":{\"C#\":1},\"summary\":\"Deterministic analysis summary\"}"
      },
      "ReviewFileChange": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/CodeLanguage"
          },
          "status": {
            "$ref": "#/components/schemas/FileChangeStatus"
          },
          "additions": {
            "type": "integer",
            "format": "int32"
          },
          "deletions": {
            "type": "integer",
            "format": "int32"
          },
          "totalChanges": {
            "type": "integer",
            "format": "int32"
          },
          "patch": {
            "type": "string",
            "nullable": true
          },
          "isBinary": {
            "type": "boolean"
          },
          "isSupported": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "ReviewFinding": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "whyItMatters": {
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/FindingSeverity"
          },
          "category": {
            "$ref": "#/components/schemas/FindingCategory"
          },
          "filePath": {
            "type": "string",
            "nullable": true
          },
          "startLine": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "endLine": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "suggestion": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewPullRequestInfo": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string"
          },
          "isDraft": {
            "type": "boolean"
          },
          "baseRef": {
            "type": "string"
          },
          "headRef": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ReviewReport": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "string"
          },
          "overallRisk": {
            "$ref": "#/components/schemas/ReviewRiskLevel"
          },
          "positiveFindings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewFinding"
            }
          },
          "recommendations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "missingTests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "securityConcerns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewFinding"
            }
          },
          "performanceConcerns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewFinding"
            }
          },
          "maintainability": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewFinding"
            }
          },
          "architectureConcerns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewFinding"
            }
          },
          "model": {
            "type": "string"
          },
          "promptTokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "completionTokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "totalTokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "duration": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false,
        "description": "Structured review report — the only AI output the rest of CodeSage consumes.",
        "example": "{\"summary\":\"Solid incremental change with one auth edge case.\",\"overallRisk\":\"Medium\",\"positiveFindings\":[\"Clear PR description\"],\"issues\":[],\"recommendations\":[\"Add a regression test for the null token path\"],\"missingTests\":[\"Auth failure path\"],\"securityConcerns\":[],\"performanceConcerns\":[],\"maintainability\":[],\"architectureConcerns\":[],\"model\":\"gpt-4o-mini\",\"promptTokens\":1200,\"completionTokens\":400,\"totalTokens\":1600,\"duration\":\"00:00:02.1500000\"}"
      },
      "ReviewRepositoryInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "fullName": {
            "type": "string"
          },
          "defaultBranch": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewRiskLevel": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "description": "Overall risk level assigned by the model.",
        "format": "int32"
      },
      "ReviewStatistics": {
        "type": "object",
        "properties": {
          "fileCount": {
            "type": "integer",
            "format": "int32"
          },
          "commitCount": {
            "type": "integer",
            "format": "int32"
          },
          "additions": {
            "type": "integer",
            "format": "int32"
          },
          "deletions": {
            "type": "integer",
            "format": "int32"
          },
          "totalChangedLines": {
            "type": "integer",
            "format": "int32"
          },
          "languagesUsed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "largestModifiedFile": {
            "type": "string",
            "nullable": true
          },
          "largestModifiedFileChanges": {
            "type": "integer",
            "format": "int32"
          },
          "testFilesChanged": {
            "type": "integer",
            "format": "int32"
          },
          "sqlFilesChanged": {
            "type": "integer",
            "format": "int32"
          },
          "configurationFilesChanged": {
            "type": "integer",
            "format": "int32"
          },
          "controllerFilesChanged": {
            "type": "integer",
            "format": "int32"
          },
          "serviceFilesChanged": {
            "type": "integer",
            "format": "int32"
          },
          "sqlModified": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}